Refactor to command archeitecture
