+++ src/main/java/org/tahomarobotics/robot/arm/ArmZeroCommand.java
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package org.tahomarobotics.robot;
+package org.tahomarobotics.robot.arm;
 
 public class ArmZeroCommand {
